fixed pytest file naming, removed update file resources.
