--- test/e2e/resources/patchbaseline_update.yaml
@@ -1,14 +1,0 @@
-apiVersion: ssm.services.k8s.aws/v1alpha1
-kind: PatchBaseline
-metadata:
-  name: test-patchbaseline
-spec:
-  name: test-patchbaseline
-  operatingSystem: WINDOWS
-  approvedPatches:
-    - KB345678
-  approvalRules:
-    patchRules:
-      - approveAfterDays: 10
-        complianceLevel: HIGH
-        enableNonSecurity: false
