+++ test/e2e/tests/test_document.py
@@ -4,23 +4,24 @@ import time
 from acktest.resources import random_suffix_name
 from acktest.k8s import resource as k8s
 from e2e.bootstrap_resources import get_bootstrap_resources
-from e2e import service_marker, load_ssm_resource
+from e2e import service_marker, load_ssm_file
 from e2e.replacement_values import REPLACEMENT_VALUES
 
-RESOURCE_PLURAL = "patchbaselines"
+RESOURCE_PLURAL = "documents"
 CREATE_WAIT_AFTER_SECONDS = 20
 DELETE_WAIT_AFTER_SECONDS = 20
 MODIFY_WAIT_AFTER_SECONDS = 20
 
+
 @pytest.fixture(scope="module")
-def patchbaseline():
-    RESOURCE_NAME = random_suffix_name("patchbaseline", 12)
+def document():
+    RESOURCE_NAME = random_suffix_name("document", 12)
 
     resources = get_bootstrap_resources()
     logging.debug(resources)
 
     replacements = REPLACEMENT_VALUES.copy()
-    resource_data = load_ssm_resource("patchbaseline", additional_replacements=replacements)
+    resource_data = load_ssm_file("document", additional_replacements=replacements,)
     reference, spec = k8s.create_custom_resource(
         resource_plural=RESOURCE_PLURAL,
         custom_resource_name=RESOURCE_NAME,
@@ -34,10 +35,11 @@ def patchbaseline():
     k8s.delete_custom_resource(reference)
     time.sleep(DELETE_WAIT_AFTER_SECONDS)
 
+
 @service_marker
-class TestPatchBaseline:
-    def test_create_delete(self, patchbaseline):
-        (reference, spec) = patchbaseline
+class TestDocument:
+    def test_create_delete(self, document):
+        (reference, spec) = document
 
         assert k8s.wait_on_condition(reference, "ACK.ResourceSynced", "True", wait_periods=10)
 
@@ -45,15 +47,30 @@ class TestPatchBaseline:
         assert cr is not None
         assert 'spec' in cr
         assert 'name' in cr["spec"]
-        assert 'operatingSystem' in cr["spec"]
-        assert 'approvedPatches' in cr["spec"]
-        assert 'approvalRules' in cr["spec"]
+        assert 'content' in cr["spec"]
 
         # Update test
-        update_data = load_ssm_resource("patchbaseline_update", additional_replacements={})
+        update_data = {
+                "spec": {
+                    "content": 
+                    {
+                        "schemaVersion": "1.2",
+                        "description": "Updated SSM Document",
+                        "mainSteps": [
+                            {
+                                "action": "aws:runShellScript",
+                                "name": "example",
+                                "inputs": {
+                                    "runCommand": ["echo Updated content"]
+                                    }
+                                }
+                            ]
+                        },
+                    },
+                }
         k8s.patch_custom_resource(reference, update_data)
         time.sleep(MODIFY_WAIT_AFTER_SECONDS)
         assert k8s.wait_on_condition(reference, "ACK.ResourceSynced", "True", wait_periods=10)
         
         updated_cr = k8s.get_resource(reference)       
-        assert updated_cr["spec"]["approvedPatches"] == update_data["spec"]["approvedPatches"]
+        assert updated_cr["spec"]["content"] == update_data["spec"]["content"]
